TouchLauncher/requirements.txt
pyxdg

# kivy deps
pygame
kivy
